# refine TaskQueue::take semantics to drain all tasks when finished