--- a/dbms/src/Flash/Pipeline/Schedule/TaskQueues/IOPriorityQueue.h
+++ b/dbms/src/Flash/Pipeline/Schedule/TaskQueues/IOPriorityQueue.h
@@ -51,6 +51,7 @@ public:
 
 private:
     void submitTaskWithoutLock(TaskPtr && task);
+
     void drainTaskQueueWithoutLock();
 
 private: