--- a/dbms/src/Flash/Pipeline/Schedule/TaskQueues/MultiLevelFeedbackQueue.cpp
+++ b/dbms/src/Flash/Pipeline/Schedule/TaskQueues/MultiLevelFeedbackQueue.cpp
@@ -176,6 +176,11 @@ bool MultiLevelFeedbackQueue<TimeGetter>::take(TaskPtr & task)
         std::unique_lock lock(mu);
         while (true)
         {
+            if (unlikely(is_finished))
+            {
+                drainTaskQueue();
+                return false;
+            }
             if (!cancel_task_queue.empty())
             {
                 task = std::move(cancel_task_queue.front());
@@ -201,8 +206,6 @@ bool MultiLevelFeedbackQueue<TimeGetter>::take(TaskPtr & task)
 
             if (queue_idx >= 0)
                 break;
-            if (unlikely(is_finished))
-                return false;
             cv.wait(lock);
         }
         level_queues[queue_idx]->take(task);
@@ -210,6 +213,28 @@ bool MultiLevelFeedbackQueue<TimeGetter>::take(TaskPtr & task)
 
     assert(task);
     return true;
+}
+
+template <typename TimeGetter>
+void MultiLevelFeedbackQueue<TimeGetter>::drainTaskQueueWithoutLock()
+{
+    while (!cancel_task_queue.empty())
+    {
+        auto task = std::move(cancel_task_queue.front());
+        cancel_task_queue.pop_front();
+        FINALIZE_TASK(task);
+    }
+
+    for (size_t i = 0; i < QUEUE_SIZE; ++i)
+    {
+        auto & cur_queue = level_queues[i];
+        TaskPtr task;
+        while (!cur_queue->empty())
+        { 
+            cur_queue->take(task);
+            FINALIZE_TASK(task);
+        }
+    }
 }
 
 template <typename TimeGetter>