--- a/dbms/src/Flash/Pipeline/Schedule/TaskQueues/MultiLevelFeedbackQueue.cpp
+++ b/dbms/src/Flash/Pipeline/Schedule/TaskQueues/MultiLevelFeedbackQueue.cpp
@@ -68,8 +68,7 @@ double UnitQueue::normalizedTimeMicrosecond()
 template <typename TimeGetter>
 MultiLevelFeedbackQueue<TimeGetter>::~MultiLevelFeedbackQueue()
 {
-    for (const auto & unit_queue : level_queues)
-        RUNTIME_ASSERT(unit_queue->empty(), logger, "all task should be taken before it is destructed");
+    drainTaskQueueWithoutLock();
 }
 
 template <typename TimeGetter>
@@ -176,11 +175,9 @@ bool MultiLevelFeedbackQueue<TimeGetter>::take(TaskPtr & task)
         std::unique_lock lock(mu);
         while (true)
         {
+            // Remaining tasks will be drained in destructor.
             if (unlikely(is_finished))
-            {
-                drainTaskQueueWithoutLock();
                 return false;
-            }
 
             if (!cancel_task_queue.empty())
             {