--- a/dbms/src/Flash/Pipeline/Schedule/TaskQueues/TaskQueue.h
+++ b/dbms/src/Flash/Pipeline/Schedule/TaskQueues/TaskQueue.h
@@ -31,7 +31,7 @@ public:
 
     virtual void submit(std::vector<TaskPtr> & tasks) = 0;
 
-    // return false if the queue is empty and finished.
+    // Will drain all tasks and return false if finished.
     virtual bool take(TaskPtr & task) = 0;
 
     // Update the execution metrics of the task taken from the queue.