--- a/dbms/src/Flash/Pipeline/Schedule/TaskQueues/tests/gtest_io_priority.cpp
+++ b/dbms/src/Flash/Pipeline/Schedule/TaskQueues/tests/gtest_io_priority.cpp
@@ -78,7 +78,10 @@ try
         {
             queue.submit(std::make_unique<MockIOTask>(context, false));
         }
-        while (!queue.empty()) { std::this_thread::sleep_for(std::chrono::milliseconds(10)); }
+        while (!queue.empty())
+        {
+            std::this_thread::sleep_for(std::chrono::milliseconds(10));
+        }
         queue.finish();
     });
     // wait