--- a/dbms/src/Flash/Pipeline/Schedule/TaskQueues/tests/gtest_mlfq.cpp
+++ b/dbms/src/Flash/Pipeline/Schedule/TaskQueues/tests/gtest_mlfq.cpp
@@ -99,7 +99,10 @@ try
             task->profile_info.addCPUExecuteTime(value);
             queue->submit(std::move(task));
         }
-        while (!queue->empty()) { std::this_thread::sleep_for(std::chrono::milliseconds(10)); }
+        while (!queue->empty())
+        {
+            std::this_thread::sleep_for(std::chrono::milliseconds(10));
+        }
         queue->finish();
     });
     // take valid task