### What problem does this PR solve?

Issue Number: close #7904

Problem Summary:

### What is changed and how it works?
This pr has no affcets on performance, check:

before this pr:
```
[2023/08/7 12:47:41 Day-219] guojiangtao-arch :: tiflash/build/master_rel ‹f2bb4c4ee*› » ./dbms/bench_dbms --benchmark_filter='MLFQBench*'
2023-08-07T12:47:41+08:00
Running ./dbms/bench_dbms
Run on (20 X 4800 MHz CPU s)
CPU Caches:
  L1 Data 48 KiB (x10)
  L1 Instruction 32 KiB (x10)
  L2 Unified 1280 KiB (x10)
  L3 Unified 25600 KiB (x1)
Load Average: 0.76, 0.43, 0.53
***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.
----------------------------------------------------------------------------
Benchmark                                  Time             CPU   Iterations
----------------------------------------------------------------------------
MLFQBench/Basic/10/1/1/10000/2    3189057026 ns      3275840 ns            1
MLFQBench/Basic/10/15/15/1000/2   4555925890 ns      5612791 ns            1
MLFQBench/Basic/10/200/200/1000/2 60053507747 ns      4526273 ns            1

----------------------------------------------------------------------------------
Benchmark                                        Time             CPU   Iterations
----------------------------------------------------------------------------------
IOPriorityBench/Basic/10/1/1/10000/2    3204130348 ns      3729885 ns            1
IOPriorityBench/Basic/10/15/15/1000/2   4550231291 ns      1301747 ns            1
IOPriorityBench/Basic/10/200/200/1000/2 60061096315 ns      4449832 ns            1
```

after this pr:
```
[2023/08/7 12:41:41 Day-219] guojiangtao-arch :: tiflash/build/master_rel ‹taskqueue_take*› » ./dbms/bench_dbms --benchmark_filter='MLFQBench*'
2023-08-07T12:42:02+08:00
Running ./dbms/bench_dbms
Run on (20 X 4800 MHz CPU s)
CPU Caches:
  L1 Data 48 KiB (x10)
  L1 Instruction 32 KiB (x10)
  L2 Unified 1280 KiB (x10)
  L3 Unified 25600 KiB (x1)
Load Average: 0.27, 0.38, 0.60
***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.
----------------------------------------------------------------------------
Benchmark                                  Time             CPU   Iterations
----------------------------------------------------------------------------
MLFQBench/Basic/10/1/1/10000/2    3189214126 ns      3435699 ns            1
MLFQBench/Basic/10/15/15/1000/2   4556525279 ns      5580156 ns            1
MLFQBench/Basic/10/200/200/1000/2 60065289566 ns      5851380 ns            1

----------------------------------------------------------------------------------
Benchmark                                        Time             CPU   Iterations
----------------------------------------------------------------------------------
IOPriorityBench/Basic/10/1/1/10000/2    3202198480 ns      3656562 ns            1
IOPriorityBench/Basic/10/15/15/1000/2   4551348436 ns      1983537 ns            1
IOPriorityBench/Basic/10/200/200/1000/2 60055732528 ns      6817641 ns            1
```
### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [ ] Integration test
- [ ] Manual test (add detailed scripts or steps below)
- [ ] No code

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- bugfix or new feature needs a release note -->

```release-note
None
```